serving_patterns/requirements/requirements_test.txt
absl-py==0.9.0
aiohttp==3.6.2
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi>=0.60.1
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gunicorn>=20.0.4
h11==0.9.0
h5py==2.10.0
httptools==0.1.1
idna==2.10
importlib-metadata>=1.7.0
joblib>=0.15.1
Keras-Preprocessing>=1.1.2
Markdown==3.2.2
mirakuru==2.3.0
more-itertools==8.4.0
numpy>=1.18.5
oauthlib==3.1.0
onnx>=1.7.0
onnxconverter-common>=1.7.0
onnxruntime>=1.4.0
opt-einsum==3.2.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
port-for==0.4
protobuf==3.12.2
psutil==5.7.0
py==1.9.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydantic>=1.5.1
pyparsing==2.4.7
pytest==5.4.3
pytest-asyncio==0.14.0
pytest-mock==3.1.1
python-multipart==0.0.5
PyYAML>=5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
skl2onnx==1.7.0
starlette>=0.13.4
tensorflow-hub==0.8.0
tensorflow-serving-api==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
torch>=1.6.0
torchvision>=0.7.0
typing>=3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
uvicorn>=0.11.7
uvloop>=0.14.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
